split codegen into two parts: 1. gen (general) filler; 2. build & exec template via filler
