+++ sql/codegen_xgboost.go
@@ -1,0 +1,9 @@
+package sql
+
+import "io"
+
+type xgboostTemplate struct {}
+
+func (*xgboostTemplate) execute(w io.Writer, r *filler) error {
+	panic("xgboostTemplate has not been implemented!")
+}
